[lbs] add safe sdmm_malloc and tagged union

+++ ft_malloc/include/sdmm.h
@@ -11,7 +11,7 @@
 
 void    	sdmm_free(void* ptr);
 void*   	sdmm_malloc(size_t size);
-void* 		sdmm_realloc(void* ptr, int size);
+void* 		sdmm_realloc(void* ptr, size_t size);
 void* 		sdmm_calloc(size_t num, size_t size);
 size_t 		sdmm_malloc_usable_size(void* ptr);
